Seperate chromap and its driver.
